src/lambda_codebase/initial_commit/bootstrap_repository/deployment/lambda_codebase/initial_commit/pipelines_repository/pipeline_types/requirements.txt
boto3
aws_cdk.aws_codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-s3
aws-cdk.aws-iam
